backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
playwright==1.48.0
google-generativeai==0.8.3
supabase==2.10.0
httpx==0.27.2
apscheduler==3.10.4
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.17
